move tests using cloud_test_catalog into func directory
