+++ tests/unit/test_listing.py
@@ -11,7 +11,6 @@ from datachain.lib.listing import (
     get_listing,
     is_listing_dataset,
     listing_uri_from_name,
-    parse_listing_uri,
 )
 from datachain.listing import Listing
 from datachain.node import DirType
@@ -179,41 +178,6 @@ def test_subtree(listing):
 def test_subdirs(listing):
     dirs = list(listing.get_dirs_by_parent_path(""))
     _match_filenames(dirs, ["dir1", "dir2"])
-
-
-@pytest.mark.parametrize(
-    "cloud_type",
-    ["s3", "azure", "gs", "file"],
-    indirect=True,
-)
-def test_parse_listing_uri(cloud_test_catalog, cloud_type):
-    ctc = cloud_test_catalog
-    catalog = ctc.catalog
-    dataset_name, listing_uri, listing_path = parse_listing_uri(
-        f"{ctc.src_uri}/dogs", catalog.cache, catalog.client_config
-    )
-    assert dataset_name == f"lst__{ctc.src_uri}/dogs/"
-    assert listing_uri == f"{ctc.src_uri}/dogs/"
-    if cloud_type == "file":
-        assert listing_path == ""
-    else:
-        assert listing_path == "dogs/"
-
-
-@pytest.mark.parametrize(
-    "cloud_type",
-    ["s3", "azure", "gs"],
-    indirect=True,
-)
-def test_parse_listing_uri_with_glob(cloud_test_catalog):
-    ctc = cloud_test_catalog
-    catalog = ctc.catalog
-    dataset_name, listing_uri, listing_path = parse_listing_uri(
-        f"{ctc.src_uri}/dogs/*", catalog.cache, catalog.client_config
-    )
-    assert dataset_name == f"lst__{ctc.src_uri}/dogs/"
-    assert listing_uri == f"{ctc.src_uri}/dogs"
-    assert listing_path == "dogs/*"
 
 
 @pytest.mark.parametrize(
